// src/app/models/flights.modals.ts
export class Flights {
  'id': string;
  'flightCode': string;
  'flightNumber': string;
  'departureDateTime': string;
  'departureStation': string;
  'arrivalDateTime': string;
  'arrivalStation': string;
  'Duration': string;
  'specialMeals': Array<string>;
  'ancillaryServices': Array<string>;
  'shoppingItems': Array<string>;
  'reservedSeats': Array<string>;
}